pi_testbench/requirements.txt
# Core dependencies
requests>=2.31.0
pyyaml>=6.0.1
pillow>=10.0.0
numpy>=1.24.0

# E-paper display
spidev>=3.6
RPi.GPIO>=0.7.1
gpiozero>=1.6.2

# Network and system monitoring
psutil>=5.9.0
netifaces>=0.11.0

# Data visualization for display
matplotlib>=3.7.0

# Async support
asyncio-mqtt>=0.13.0
aiohttp>=3.8.0

# Logging
colorlog>=6.7.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0 